REFACTOR: Usando funcionalidades já existentes dos modulos para remover duplicações (#32)

+++ src/modules/uploader/localStorage.ts
@@ -12,14 +12,10 @@ export class LocalStorageProvider implements StorageProvider {
   async save({ filename, folder, file }: SaveFileParams): Promise<SavedFile> {
     const uploadFolder = path.join("uploads", folder);
 
-    if (!fs.existsSync(uploadFolder)) {
-      fs.mkdirSync(uploadFolder, {
-        recursive: true,
-      });
-    }
+    if (!fs.existsSync(uploadFolder))
+      fs.mkdirSync(uploadFolder, { recursive: true });
 
     const filepath = path.join(uploadFolder, filename);
-
     await pipeline(file, fs.createWriteStream(filepath));
 
     return {
@@ -29,8 +25,6 @@ export class LocalStorageProvider implements StorageProvider {
   }
 
   async delete(filepath: string): Promise<void> {
-    if (fs.existsSync(filepath)) {
-      await fs.promises.unlink(filepath);
-    }
+    if (fs.existsSync(filepath)) await fs.promises.unlink(filepath);
   }
 }
